Price Calculation

+++ Dev/Code/Source/Alukkas/Libraries/Orbio.Core/Domain/Catalog/Product.cs
@@ -1,5 +1,6 @@
 ﻿using System.Runtime.Serialization;
 using Orbio.Core.Domain.Catalog.Abstract;
+using System.Collections.Generic;
 
 namespace Orbio.Core.Domain.Catalog
 {
@@ -76,7 +77,7 @@ namespace Orbio.Core.Domain.Catalog
         /// Gets or sets ProductUnit
         /// </summary>
         [DataMember]
-        public int ProductUnit { get; set; }
+        public decimal ProductUnit { get; set; }
 
         /// <summary>
         /// Gets or sets PriceUnit
